Factor PayeeMenu out to component file
